+++ Client/src/types.d.ts
@@ -24,3 +24,11 @@ export interface CardProps2 extends CardProps {
   description: string;
   image: string[];
 }
+
+export interface Comment {
+    id: number;
+    userId: string;
+    productId: number;
+    Comment: string;
+    date: string;
+  }
